+++ common/src/python/curator/scheduling.py
@@ -55,11 +55,13 @@ def build_file_heap(subject: Subject) -> MinHeap[FileModel]:
     try:
         response_model.associate_uds_session()
     except ValueError as error:
-        log.warning(f"{subject.label}: {error}")
+        log.error(f"{subject.label} failed, clearing curation: {error}")
         # clear out curation tags on all files
         for file_model in response_model.data:
             curator.clear_curation_tag(file_model)
 
+        # write error
+        curator.handle_curation_failure(subject, str(error))
         return heap
 
     log.debug("Curating %s files in for %s", len(response_model.data), subject.label)
@@ -89,7 +91,7 @@ def curate_subject(subject_id: str) -> None:
 
     heap = build_file_heap(subject)
     if not heap:
-        log.warning(f"No files to curate for subject {subject_id}")
+        log.warning(f"No files to curate for subject {subject.label}")
         return
 
     log.debug(f"Curating {len(heap)} files for {subject.label}")