+++ common/src/python/curator/scheduling_models.py
@@ -23,7 +23,7 @@ VISIT_PASS_LITERALS = Literal["pass0", "pass1", "pass2", "pass3"]
 VISIT_PATTERN = re.compile(
     r"^"
     r"(?P<pass2>.+("
-    r"_BDS|_CSF|_NP|_MDS|_MLST|_MEDS|_FTLD|_LBD|"
+    r"_BDS|_CSF|_NP|_MDS|_MLST|_MEDS|_FTLD|_LBD|_B1A|"
     r"apoe_genotype|NCRAD-SAMPLES.+|niagads_availability|"
     r"SCAN-MR-QC.+|SCAN-MR-SBM.+|"
     r"SCAN-PET-QC.+|SCAN-AMYLOID-PET-GAAIN.+|SCAN-AMYLOID-PET-NPDKA.+|"
@@ -61,6 +61,7 @@ SCOPE_PATTERN = re.compile(
     r"(?P<meds>.+_MEDS\.json)|"
     r"(?P<ftld>.+_FTLD\.json)|"
     r"(?P<lbd>.+_LBD\.json)|"
+    r"(?P<b1a>.+_B1A\.json)|"
     r"(?P<uds>.+_UDS\.json)"
     r"$"
 )